pyproject.toml
[tool.poetry]
name = "SK42"
version = "0.1.0"
description = ""
authors = ["vlakir <vladimir@kirievskiy.ru>"]
readme = "README.md"

 package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.15"
pyproj = "^3.7.2"
pydantic = "^2.9.0"
python-dotenv = "^1.1.1"
ruff = "^0.12.10"
mypy = "^1.17.1"
tomlkit = "^0.13.3"
aiohttp = "^3.12.15"
aiohttp-client-cache = "^0.11.1"
aiosqlite = "^0.21.0"
pyside6 = "^6.9.1"
psutil = "^6.0.0"
pyinstaller = "^6.15.0"
pillow = "^11.3.0"
numpy = "^2.2.0"
numba = "^0.61.0"
scipy = "^1.15.0"
opencv-python = "^4.10.0"
pytest = "^8.4.2"
certifi = "^2025.10.5"
pytest-cov = "^7.0.0"
pytest-asyncio = "^1.3.0"
superqt = "^0.7.8"
markdown-it-py = "^3.0.0"
mdit-py-plugins = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]

unsafe-fixes = true

line-length = 88
indent-width = 4
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "business-facing/layer",
    "migrations",
    "draft",
    "tests",
]

target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [ "D1", "D203", "D212",
    "BLE001", "ANN002", "ANN204",
    "RUF001", "COM812", "RUF002",
    "RUF003", "G004", "PLR0912",
    "PLR0913", "D401", "PLR0915",
    "C901", "D205", "N802",
    # Lazy imports are intentional architectural pattern
    # (avoid circular imports, defer heavy modules)
    "PLC0415"]

[tool.ruff.lint.per-file-ignores]
# Matrix rain animation uses random for visual effects, not cryptography
"src/gui/matrix_rain.py" = ["S311"]

fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 72
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.mypy]
explicit_package_bases = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-q --tb=short --cov=src --cov-report=term-missing -m 'not integration'"
pythonpath = ["src"]
markers = [
    "integration: Integration tests requiring Mapbox API key and network access",
]
filterwarnings = [
    "ignore::DeprecationWarning:pydantic.*:",
    "ignore::ResourceWarning:aiohttp.*:",
]

[tool.coverage.run]
omit = ["src/gui/*", "src/gui/**"]


[tool.black]
line-length = 88

[tool.autopep8]
max_line_length = 88

[tool.poetry.scripts]
helmert_tool = "gui.helmert_tool:main"


[dependency-groups]
dev = [
    "pytest-cov (>=7.0.0,<8.0.0)",
]
